Pkg.generate: generate missing files when not forced.
This is especially useful for generating license files for packages
that don't already have them.

+++ base/pkg/generate.jl
@@ -12,44 +12,46 @@ function package(
     force::Bool = false,
     authors::String = copyright_name(),
     years::Union(Int,String) = copyright_year(),
-    username::String = github_user(),
+    user::String = github_user(),
 )
     isnew = !ispath(pkg)
-    if !isnew
-        force || error("$pkg exists, refusing to overwrite.")
-        Git.dirty(dir=pkg) && error("$pkg is dirty – commit or stash your changes")
-    end
-    url = isempty(username) ? "" : "git://github.com/$username/$pkg.jl.git"
     try
-        Generate.init(pkg,url)
-        Generate.license(pkg,license,years,authors)
-        Generate.readme(pkg,username)
-        Generate.entrypoint(pkg)
-        Generate.travis(pkg)
-
-        msg = """
-        $pkg.jl $(isnew ? "generated" : "regenerated") files.
-
-            license:  $license
-            authors:  $authors
-            years:    $years
-            username: $username
-
-        Julia Version $VERSION [$(Base.BUILD_INFO.commit[1:10])]
-        """
-
         if isnew
-            info("Committing $pkg generated files")
-            Git.run(`commit -q -m $msg`, dir=pkg)
-        elseif Git.staged(dir=pkg)
-            Git.run(`reset -q --`, dir=pkg)
-            info("Regenerated files left unstaged, use `git add -p` to select")
-            open(io->print(io,msg), joinpath(Git.dir(pkg),"MERGE_MSG"), "w")
+            url = isempty(user) ? "" : "git://github.com/$user/$pkg.jl.git"
+            Generate.init(pkg,url)
         else
-            info("Regenerated files are unchanged")
+            Git.dirty(dir=pkg) && error("$pkg is dirty – commit or stash your changes")
         end
+        # Git.transact(dir=pkg) do
+            Generate.license(pkg,license,years,authors,force=force)
+            Generate.readme(pkg,user,force=force)
+            Generate.entrypoint(pkg,force=force)
+            Generate.travis(pkg,force=force)
+
+            msg = """
+            $pkg.jl $(isnew ? "generated" : "regenerated") files.
+
+                license:  $license
+                authors:  $authors
+                years:    $years
+                user:     $user
+
+            Julia Version $VERSION [$(Base.BUILD_INFO.commit[1:10])]
+            """
+
+            if isnew
+                info("Committing $pkg generated files")
+                Git.run(`commit -q -m $msg`, dir=pkg)
+            elseif Git.dirty(dir=pkg)
+                Git.run(`reset -q --`, dir=pkg)
+                info("Regenerated files left unstaged, use `git add -p` to select")
+                open(io->print(io,msg), joinpath(Git.dir(pkg),"MERGE_MSG"), "w")
+            else
+                info("Regenerated files are unchanged")
+            end
+        # end
     catch
-        isnew ? run(`rm -rf $pkg`) : Git.run(`checkout -q -f`, dir=pkg)
+        isnew && run(`rm -rf $pkg`)
         rethrow()
     end
 end
@@ -65,27 +67,28 @@ function init(pkg::String, url::String="")
     Git.set_remote_url(url,dir=pkg)
 end
 
-function license(pkg::String, license::String, years::Union(Int,String), authors::String)
-    if !haskey(LICENSES,license)
-        licenses = join(sort!([keys(LICENSES)...], by=lowercase), ", ")
-        error("$license is not a known license choice, choose one of: $licenses.")
-    end
-    genfile(pkg,"LICENSE.md") do io
+function license(pkg::String, license::String,
+                 years::Union(Int,String), authors::String; force::Bool=false)
+    genfile(pkg,"LICENSE.md",force) do io
+        if !haskey(LICENSES,license)
+            licenses = join(sort!([keys(LICENSES)...], by=lowercase), ", ")
+            error("$license is not a known license choice, choose one of: $licenses.")
+        end
         print(io, LICENSES[license](pkg, string(years), authors))
-    end
+    end || info("License file exists, leaving unmodified; use `force=true` to overwrite")
 end
 
-function readme(pkg::String, username::String="")
-    genfile(pkg,"README.md") do io
+function readme(pkg::String, user::String=""; force::Bool=false)
+    genfile(pkg,"README.md",force) do io
         println(io, "# $pkg")
-        isempty(username) && return
-        url = "https://travis-ci.org/$username/$pkg.jl"
+        isempty(user) && return
+        url = "https://travis-ci.org/$user/$pkg.jl"
         println(io, "\n[![Build Status]($url.png)]($url)")
     end
 end
 
-function travis(pkg::String)
-    genfile(pkg,".travis.yml") do io
+function travis(pkg::String; force::Bool=false)
+    genfile(pkg,".travis.yml",force) do io
         print(io, """
         language: cpp
         compiler:
@@ -104,8 +107,8 @@ function travis(pkg::String)
     end
 end
 
-function entrypoint(pkg::String)
-    genfile(pkg,"src/$pkg.jl") do io
+function entrypoint(pkg::String; force::Bool=false)
+    genfile(pkg,"src/$pkg.jl",force) do io
         print(io, """
         module $pkg
 
@@ -116,12 +119,16 @@ function entrypoint(pkg::String)
     end
 end
 
-function genfile(f::Function, pkg::String, file::String)
+function genfile(f::Function, pkg::String, file::String, force::Bool=false)
     path = joinpath(pkg,file)
-    info("Generating $file")
-    mkpath(dirname(path))
-    open(f, path, "w")
-    Git.run(`add $file`, dir=pkg)
+    if force || !ispath(path)
+        info("Generating $file")
+        mkpath(dirname(path))
+        open(f, path, "w")
+        Git.run(`add $file`, dir=pkg)
+        return true
+    end
+    return false
 end
 
 mit(pkg::String, years::String, authors::String) =
